+++ apps/framework-cli/src/framework/typescript.rs
@@ -1,5 +1,6 @@
 pub mod aggregation;
 pub mod bin;
+pub mod checker;
 pub mod consumption;
 pub mod export_collectors;
 pub mod generator;
