+++ apps/framework-cli/src/framework/typescript/checker.rs
@@ -1,0 +1,11 @@
+use crate::framework::core::check::{CheckerError, SystemChecker};
+use async_trait::async_trait;
+
+pub struct TypeScriptChecker;
+
+#[async_trait]
+impl SystemChecker for TypeScriptChecker {
+    async fn validate(&self) -> Result<(), CheckerError> {
+        Ok(())
+    }
+}
